Move the implementaion into a namespace and a class
This allows for trivial autoloading of the requirements along with not poluting the global namespace with functioms.

While the polution can be avoided by specifying just the namespace it will not resolve the autoloading issue. Where it is possible to directly invoke a function herein and have it not be loaded yet thus causing an error. Better to wrap it in a class in php.