Resolve phpstan issues
Phpstorm and phpstan had a lot of issues with the file.

I have specified return types, parameter types, explicit type conversions and such.

The variance in the "mode" parameter's return value is resolved by using a match statement. While a default option for returning '' could be included the methods are now private and as such should not be invoked anyway.